+++ app/controllers/v2/posts_controller.rb
@@ -1,0 +1,8 @@
+module V2
+  class PostsController < ApplicationController
+    # GET / posts
+    def index
+      json_response({ message: "" })
+    end
+  end
+end
