+++ spec/controllers/v2/posts_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe V2::PostsController, type: :controller do
+
+end
